Renaming GetContactDetailsResult -> GetContactDetailsResponse, and moved to a service-local 'Types' namespace.
