+++ src/SFA.DAS.LevyTransferMatching.Infrastructure/Services/OpportunitiesService/IOpportunitiesService.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using SFA.DAS.LevyTransferMatching.Infrastructure.Dto;
-using SFA.DAS.LevyTransferMatching.Infrastructure.Models;
+using SFA.DAS.LevyTransferMatching.Infrastructure.Services.OpportunitiesService.Types;
 
 namespace SFA.DAS.LevyTransferMatching.Infrastructure.Services.OpportunitiesService
 {
@@ -11,6 +11,6 @@ namespace SFA.DAS.LevyTransferMatching.Infrastructure.Services.OpportunitiesServ
 
         Task<OpportunityDto> GetOpportunity(int id);
 
-        Task<GetContactDetailsResult> GetContactDetails(long accountId, int pledgeId);
+        Task<GetContactDetailsResponse> GetContactDetails(long accountId, int pledgeId);
     }
 }
